+++ lib/magic/presenter/base.rb
@@ -13,12 +13,18 @@ module Magic
 		#
 		# Presenters provide automatic class inference for any model based
 		# on its class name powered by Magic Lookup.
-		# For example, `MyModel.new` looks for `MyModelPresenter` first.
+		# For example, `MyModel.new` looks for `MyPresenter` first.
 		# If not found, it looks for presenters of its ancestor classes,
 		# up to `ObjectPresenter`.
 		class Base < Decorator::Base
 			class << self
-				def name_for(object_class) = "#{object_class}Presenter"
+				def name_for object_class
+					object_class
+							.to_s
+							.delete_suffix('Model')
+							.delete_suffix('Record')
+							.then { "#{_1}Presenter" }
+				end
 			end
 		end
 	end