+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - Lookup error reporting: `Magic::Presenter` should be used as a lookup base instead of `Magic::Decorator`.
+- Magic Lookup requires presenters to be eagerly loaded to be accessible via lookups.
 
 
 ## [0.1.0] — 2024-10-19